requirements.txt
tensorflow
numpy
pillow
matplotlib
seaborn
scikit-learn
flask
flask-cors
werkzeug
requests
tqdm
